Auto-adding duration works

+++ app/models/recording.rb
@@ -6,7 +6,7 @@ class Recording < ApplicationRecord
 
   attr_encrypted :audio, key: Rails.application.config.audio_encryption_key, encode: false, encode_iv: false
 
-  after_save :set_duration
+  after_validation :set_duration
   TMP_PATH = "#{Rails.root}/recordings_tmp"
 
   def path
@@ -15,7 +15,7 @@ class Recording < ApplicationRecord
 
   # Creates tmp file for recording and returns file size in bytes
   def make_tmp_file!
-    File.open(path, 'wb') { |file| file.write(self.audio) }
+    File.open(path, 'wb') { |file| file.write(self.audio) } unless File.exists? path
   end
 
   def tmp_file_size
@@ -26,7 +26,7 @@ class Recording < ApplicationRecord
 
   def set_duration
     make_tmp_file!
-    self.update! duration: (tmp_file_size / 10332)
+    self.duration = (tmp_file_size / 10332)
   end
 
 end
